# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Elevator Inspection Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /elevators:
    get:
      operationId: listElevators
      tags: [elevators]
      summary: List all elevators
      security:
        - bearerAuth: []
      parameters:
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [traction, hydraulic, other]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of elevators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Elevator"
    post:
      operationId: createElevator
      tags: [elevators]
      summary: Create an elevator
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateElevatorBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Elevator"
  /elevators/{id}:
    get:
      operationId: getElevator
      tags: [elevators]
      summary: Get an elevator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Elevator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Elevator"
    put:
      operationId: updateElevator
      tags: [elevators]
      summary: Update an elevator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateElevatorBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Elevator"
    delete:
      operationId: deleteElevator
      tags: [elevators]
      summary: Delete an elevator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /inspections:
    get:
      operationId: listInspections
      tags: [inspections]
      summary: List all inspections
      security:
        - bearerAuth: []
      parameters:
        - name: elevatorId
          in: query
          schema:
            type: integer
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [NOT_STARTED, SCHEDULED, IN_PROGRESS, COMPLETED, OVERDUE]
        - name: inspectionType
          in: query
          schema:
            type: string
            enum: [CAT1, CAT5]
        - name: elevatorType
          in: query
          schema:
            type: string
            enum: [traction, hydraulic, other]
        - name: search
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of inspections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inspection"
    post:
      operationId: createInspection
      tags: [inspections]
      summary: Create an inspection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInspectionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
  /inspections/{id}:
    get:
      operationId: getInspection
      tags: [inspections]
      summary: Get an inspection
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inspection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
    put:
      operationId: updateInspection
      tags: [inspections]
      summary: Update an inspection
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInspectionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
    delete:
      operationId: deleteInspection
      tags: [inspections]
      summary: Delete an inspection
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /export/inspections:
    get:
      operationId: exportInspections
      tags: [export]
      summary: Export inspections to Excel
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: inspectionType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /export/elevators:
    get:
      operationId: exportElevators
      tags: [export]
      summary: Export elevators to Excel
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: buildingId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Elevator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        bank:
          type: string
        type:
          type: string
          enum: [traction, hydraulic, other]
        buildingId:
          type: integer
        buildingName:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        organizationId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - buildingId
        - organizationId
        - createdAt
    CreateElevatorBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        bank:
          type: string
        type:
          type: string
          enum: [traction, hydraulic, other]
        buildingId:
          type: integer
      required:
        - name
        - type
        - buildingId
    Inspection:
      type: object
      properties:
        id:
          type: integer
        elevatorId:
          type: integer
        elevatorName:
          type: string
        buildingId:
          type: integer
        buildingName:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        inspectionType:
          type: string
          enum: [CAT1, CAT5]
        recurrenceYears:
          type: integer
        lastInspectionDate:
          type: string
          format: date
        nextDueDate:
          type: string
          format: date
        scheduledDate:
          type: string
          format: date
        completionDate:
          type: string
          format: date
        status:
          type: string
          enum: [NOT_STARTED, SCHEDULED, IN_PROGRESS, COMPLETED, OVERDUE]
        notes:
          type: string
        organizationId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - elevatorId
        - inspectionType
        - recurrenceYears
        - status
        - organizationId
        - createdAt
    CreateInspectionBody:
      type: object
      properties:
        elevatorId:
          type: integer
        inspectionType:
          type: string
          enum: [CAT1, CAT5]
        recurrenceYears:
          type: integer
        lastInspectionDate:
          type: string
          format: date
        scheduledDate:
          type: string
          format: date
        completionDate:
          type: string
          format: date
        status:
          type: string
          enum: [NOT_STARTED, SCHEDULED, IN_PROGRESS, COMPLETED, OVERDUE]
        notes:
          type: string
      required:
        - elevatorId
        - inspectionType
        - recurrenceYears
...