# lib/api-zod/src/generated/types/createElevatorBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Elevator Inspection Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateElevatorBodyType } from "./createElevatorBodyType";

export interface CreateElevatorBody {
  name: string;
  description?: string;
  bank?: string;
  type: CreateElevatorBodyType;
  buildingId: number;
}